Update for unreleased GP API 0.2
- Updated to GP API v0.2
- Added `/isa transfer` command to transfer island ownership
- Island claims no longer use a player's claim blocks
- Using `/deleteclaim` on an island claim requires confirmation to delete the island
- Island spawns are now set with the player's current rotation (Not saved yet)
- `/isa delete <user> [regen]` now accepts a optional boolean to disable region regeneration

+++ src/main/java/net/mohron/skyclaims/database/SqliteDatabase.java
@@ -1,6 +1,6 @@
 package net.mohron.skyclaims.database;
 
-import com.flowpowered.math.vector.Vector3i;
+import com.flowpowered.math.vector.Vector3d;
 import com.google.common.collect.Maps;
 import net.mohron.skyclaims.SkyClaims;
 import net.mohron.skyclaims.config.type.SqliteConfig;
@@ -14,9 +14,7 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.Collection;
 import java.util.HashMap;
-import java.util.Map;
 import java.util.UUID;
 
 public class SqliteDatabase extends Database {
@@ -143,7 +141,7 @@ public class SqliteDatabase extends Database {
 				int z = results.getInt("spawnZ");
 				boolean locked = results.getBoolean("locked");
 
-				Vector3i spawnLocation = new Vector3i(x, y, z);
+				Vector3d spawnLocation = new Vector3d(x, y, z);
 //				SkyClaims.getInstance().getLogger().debug(String.format("Loading %s, %s, %s, %s, %s", islandId, ownerId, claimId, spawnLocation.toString(), locked));
 				Island island = new Island(islandId, ownerId, claimId, spawnLocation, locked);
 
@@ -177,7 +175,7 @@ public class SqliteDatabase extends Database {
 				int z = results.getInt("z");
 
 				UUID id = UUID.randomUUID();
-				Vector3i spawnLocation = new Vector3i(x, y, z);
+				Vector3d spawnLocation = new Vector3d(x, y, z);
 				Island island = new Island(id, ownerId, claimId, spawnLocation, false);
 
 				islands.put(id, island);
